requirements.txt
Flask
Flask-Login
Flask-Assets
jsmin
cssmin
Flask-MongoEngine